Remove `authorAvatar` from 7-Dec-2015 post
The gravatar image is pulled from the author's email address. I
originally added the `authorAvatar` to try and fix the absence of my
avatar image from the post, but that was the wrong fix. Instead, I
simply added my author email address to my gravatar profile and that
fixed the problem.